renamed trip and route classes
