--- app/src/main/java/com/example/ithaca_transit_android_v2/Models/ActionType.kt
@@ -1,12 +1,0 @@
-package com.example.ithaca_transit_android_v2.Models
-
-/*
-Represents the types of action a user can take while on their Trip.
- */
-enum class ActionType {
-    ARRIVE,
-    DEPART,
-    GET_OFF_BUS,
-    BOARD,
-    STAY_ON_BUS
-}
