+++ app/src/main/java/com/example/ithaca_transit_android_v2/models/ActionType.kt
@@ -1,0 +1,12 @@
+package com.example.ithaca_transit_android_v2.models
+
+/*
+Represents the types of action a user can take while on their Route.
+ */
+enum class ActionType {
+    ARRIVE,
+    DEPART,
+    GET_OFF_BUS,
+    BOARD,
+    STAY_ON_BUS
+}
