+++ app/src/main/java/com/example/ithaca_transit_android_v2/states/RouteCardState.kt
@@ -1,28 +1,28 @@
 package com.example.ithaca_transit_android_v2.states
 
-import com.example.ithaca_transit_android_v2.Models.Trip
+import com.example.ithaca_transit_android_v2.models.Route
 
 // Parent Class
 sealed class RouteCardState (
-    val trips: List<Trip> = listOf()
+    val routes: List<Route> = listOf()
 )
 
 /*
  * A default display of a route card in route option view.
- * [trip.isWalkingOnly] and number of [trip.routes] affects the drawing of transit details.
+ * [route.isWalkingOnly] and number of [route.directions] affects the drawing of transit details.
  */
-class RouteOptionState(trips: List<Trip>): RouteCardState(trips = trips)
+class RouteOptionState(routes: List<Route>): RouteCardState(routes = routes)
 
 /*
  * A route card from [RouteOptionState] is clicked.
- * [RouteDetailViewState] displays the selected trip in detail.
- * Takes in only one [trip] object because only one trip is displayed.
+ * [RouteDetailViewState] displays the selected route in detail.
+ * Takes in only one [route] object because only one route is displayed.
  */
-class RouteDetailViewState(trip: Trip): RouteCardState(listOf(trip))
+class RouteDetailViewState(route: Route): RouteCardState(listOf(route))
 
 /*
  * User clicks on the [Leave Now] button in [RouteOptionState] and a dialog pops up (UI change).
  * When the dialog box is set, do a networking call and go back to [RouteOptionState] with
- * the routes refreshed.
+ * the directions refreshed.
  */
 class LeaveAtClickedState: RouteCardState()
